pyproject.toml
[project]
name = "tgbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=22.3",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.3",
    "tabulate>=0.9.0",
    "minio>=7.2.0",
    "transformers>=4.54.1",
]

[[tool.uv.index]]
name = "jetson-jp6-cu126"
url = "https://pypi.jetson-ai-lab.io/jp6/cu126/+simple"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "I"]
ignore = ["E501"]
unfixable = ["B"]
